Add token.Remove, move test to multitenancy test file
